refactor(ecs): move accessors to system dir
