+++ core/include/cubos/core/ecs/system/system.hpp
@@ -12,9 +12,9 @@
 #include <typeindex>
 #include <unordered_set>
 
-#include <cubos/core/ecs/accessors.hpp>
 #include <cubos/core/ecs/event_reader.hpp>
 #include <cubos/core/ecs/event_writer.hpp>
+#include <cubos/core/ecs/system/accessors.hpp>
 #include <cubos/core/ecs/system/commands.hpp>
 #include <cubos/core/ecs/system/query.hpp>
 #include <cubos/core/ecs/world.hpp>
